delete some stuff that was maybe breaking the editor
this commit removes the transition_control addon, which was kind of an
experiment that didn't actually work (i think) and anyway we never ended
up using, plus the output_font BitmapFont resource.  the culprit was
actually the bitmap font iirc -  bitmap fonts are fucked in all sorts of
ways since godot 3 (see docs/bitmap_notes.md); idk if godot 4 fixed them
but it's fine, we've been using a ttf since then instead and it works.