pufine anims (wip)
adding front & back idles for pufine (the "elegant" evolution of pufig),
plus a half-finished front walk.  i think these look pretty nice but
gosh do they take forever!! i'm hoping to finish all the animations for
this silly sheep-pig by the end of the week (after back-walk, we just
need front and back versions of sleep and pre-sleep).

we also don't have a datafile for pufine yet.  actually, instead of
making it manually, i'd like to build some simple tooling to edit
datafiles in the godot editor. (my pipe dream is to be able to do this
in the game itself, or at least in a separate bundled executable, but an
editor plugin would be a good POC)